Refactor our impl_macros
