+++ crates/module-system/sov-modules-api/src/gas/mod.rs
@@ -1,3 +1,4 @@
+mod impl_macros;
 mod metered_utils;
 mod meters;
 mod price;
@@ -7,6 +8,7 @@ mod unit;
 #[cfg(test)]
 mod tests;
 
+pub(crate) use impl_macros::*;
 pub use metered_utils::{
     charge_gas_to_deserialize_json, metered_credential, MeteredBorshDeserialize,
     MeteredBorshDeserializeError, MeteredHasher, MeteredSigVerificationError, MeteredSignature,
